+++ torch_geometric/backend.py
@@ -23,8 +23,8 @@ def use_segment_matmul_heuristic(
     :meth:`segment_matmul` can speed up computation.
     """
     # NOTE This heuristic was learned on an A100 via sklearn using a simple
-    # StandardScaler() -> LinearSVC() model. For now, it is only used
-    # in combination with RGCNConv.
+    # StandardScaler() -> LinearSVC() model.
+    # For now, it is only used in combination with `RGCNConv`.
     x = torch.tensor([
         num_segments,
         max_segment_size,