+++ torch_geometric/nn/dense/linear.py
@@ -2,7 +2,7 @@ import copy
 import math
 import sys
 import time
-from typing import Any, Dict, Optional, Union
+from typing import Any, Dict, Optional, Tuple, Union
 
 import torch
 import torch.nn.functional as F
@@ -16,11 +16,6 @@ from torch_geometric.nn import inits
 from torch_geometric.typing import pyg_lib
 from torch_geometric.utils import index_sort
 from torch_geometric.utils.sparse import index2ptr
-
-if "pytest" in sys.modules:
-    MEASURE_ITER = 1
-else:
-    MEASURE_ITER = 3
 
 
 def is_uninitialized_parameter(x: Any) -> bool:
@@ -243,8 +238,6 @@ class HeteroLinear(torch.nn.Module):
         self.is_sorted = is_sorted
         self.kwargs = kwargs
 
-        self._timing_cache = {}
-
         if self.in_channels == -1:
             self.weight = torch.nn.parameter.UninitializedParameter()
             self._hook = self.register_forward_pre_hook(
@@ -258,6 +251,9 @@ class HeteroLinear(torch.nn.Module):
             self.register_parameter('bias', None)
         self.reset_parameters()
 
+        # Timing cache for benchmarking naive vs. segment matmul usage:
+        self._timing_cache: Dict[int, Tuple[float, float]] = {}
+
     def reset_parameters(self):
         r"""Resets all learnable parameters of the module."""
         reset_weight_(self.weight, self.in_channels,
@@ -265,100 +261,78 @@ class HeteroLinear(torch.nn.Module):
         reset_bias_(self.bias, self.in_channels,
                     self.kwargs.get('bias_initializer', None))
 
-    @torch.jit.unused
-    def forward_segmm(self, x: Tensor, type_vec_ptr: Tensor) -> Tensor:
-        assert self.weight is not None
-        out = pyg_lib.ops.segment_matmul(x, type_vec_ptr, self.weight)
-        return out
-
-    def forward_naive(self, x: Tensor, type_vec_ptr: Tensor) -> Tensor:
+    def forward_naive(self, x: Tensor, type_ptr: Tensor) -> Tensor:
         out = x.new_empty(x.size(0), self.out_channels)
-
-        for i in range(self.num_types):
-            off_start, off_end = type_vec_ptr[i], type_vec_ptr[i + 1]
-            subset_out = x[off_start:off_end] @ self.weight[i]
-            # The data type may have changed with mixed precision:
-            out[off_start:off_end] = subset_out.to(out.dtype)
-
+        for i, (start, end) in enumerate(zip(type_ptr[:-1], type_ptr[1:])):
+            out[start:end] = x[start:end] @ self.weight[i]
         return out
 
-    @torch.jit.unused
-    def _update_timing_cache(self, x: Tensor, type_vec_ptr: Tensor,
-                             num_rows: int) -> bool:
-        with torch.no_grad():
-            # only measure forward pass for now
-            if torch.cuda.is_available():
-                torch.cuda.synchronize()
-            start = time.perf_counter()
-            for _ in range(MEASURE_ITER):
-                _ = self.forward_segmm(x, type_vec_ptr)
-            if torch.cuda.is_available():
-                torch.cuda.synchronize()
-            end = time.perf_counter()
-            time_segmm = end - start
+    def forward_segmm(self, x: Tensor, type_ptr: Tensor) -> Tensor:
+        return pyg_lib.ops.segment_matmul(x, type_ptr, self.weight)
 
-            if torch.cuda.is_available():
-                torch.cuda.synchronize()
-            start = time.perf_counter()
-            for _ in range(MEASURE_ITER):
-                _ = self.forward_naive(x, type_vec_ptr)
-            if torch.cuda.is_available():
-                torch.cuda.synchronize()
-            end = time.perf_counter()
-            time_naive = end - start
+    @torch.no_grad()
+    def _update_timing_cache(
+        self,
+        x: Tensor,
+        type_ptr: Tensor,
+        key: int,
+    ) -> None:
 
-            # first entry is with segmm, second without
-            # if segmm is faster based on timings, use it
-            self._timing_cache[num_rows] = (time_segmm, time_naive)
-            use_segment_matmul = time_segmm < time_naive
+        MEASURE_ITER = 1 if 'pytest' in sys.modules else 3
 
-        return use_segment_matmul
+        if torch.cuda.is_available():
+            torch.cuda.synchronize()
+        t = time.perf_counter()
+        for _ in range(MEASURE_ITER):
+            _ = self.forward_segmm(x, type_ptr)
+        if torch.cuda.is_available():
+            torch.cuda.synchronize()
+        time_segmm = time.perf_counter() - t
+
+        if torch.cuda.is_available():
+            torch.cuda.synchronize()
+        t = time.perf_counter()
+        for _ in range(MEASURE_ITER):
+            _ = self.forward_naive(x, type_ptr)
+        if torch.cuda.is_available():
+            torch.cuda.synchronize()
+        time_naive = time.perf_counter() - t
+
+        self._timing_cache[key] = (time_segmm, time_naive)
 
     def forward(self, x: Tensor, type_vec: Tensor) -> Tensor:
-        r"""Forward pass.
+        r"""The forward pass.
 
         Args:
             x (torch.Tensor): The input features.
             type_vec (torch.Tensor): A vector that maps each entry to a type.
         """
         perm: Optional[Tensor] = None
-        if not self.is_sorted:
-            if (type_vec[1:] < type_vec[:-1]).any():
-                type_vec, perm = index_sort(type_vec, self.num_types)
-                x = x[perm]
+        if not self.is_sorted and (type_vec[1:] < type_vec[:-1]).any():
+            type_vec, perm = index_sort(type_vec, self.num_types)
+            x = x[perm]
 
-        type_vec_ptr = index2ptr(type_vec, self.num_types)
+        type_ptr = index2ptr(type_vec, self.num_types)
 
         if torch_geometric.backend.use_segment_matmul is None:
             use_segment_matmul = False
+            if (torch_geometric.typing.WITH_SEGMM and not is_compiling()
+                    and not torch.jit.is_scripting()):
 
-            # TODO check cases of compiling and scripting properly
-            if torch_geometric.typing.WITH_SEGMM and not is_compiling(
-            ) and not torch.jit.is_scripting():
-                # to avoid too many measurements for dynamic shapes
-                # use "magnitude" of number of rows as target
-                num_rows = math.floor(math.log10(x.size(0)))
-                if num_rows in self._timing_cache:
-                    timings = self._timing_cache[num_rows]
-                    # first entry is with segmm, second without
-                    # if segmm is faster based on timings, use it
-                    use_segment_matmul = timings[0] < timings[1]
-
-                elif num_rows not in self._timing_cache:
-                    use_segment_matmul = self._update_timing_cache(
-                        x, type_vec_ptr, num_rows)
-
+                # Use "magnitude" of number of rows as timing key:
+                key = math.floor(math.log10(x.size(0)))
+                if key not in self._timing_cache:
+                    self._update_timing_cache(x, type_ptr, key)
+                time_segmm, time_naive = self._timing_cache[key]
+                use_segment_matmul = time_segmm < time_naive
         else:
-            # TODO check cases of compiling and scripting properly
-            use_segment_matmul = (torch_geometric.typing.WITH_SEGMM and
-                                  torch_geometric.backend.use_segment_matmul
-                                  and not is_compiling()
-                                  and not torch.jit.is_scripting())
+            use_segment_matmul = torch_geometric.backend.use_segment_matmul
 
-        if use_segment_matmul:
-            out = self.forward_segmm(x, type_vec_ptr)
+        if (torch_geometric.typing.WITH_SEGMM and not is_compiling()
+                and use_segment_matmul):
+            out = self.forward_segmm(x, type_ptr)
         else:
-            out = self.forward_naive(x, type_vec_ptr)
+            out = self.forward_naive(x, type_ptr)
 
         if self.bias is not None:
             out += self.bias[type_vec]